:lipstick: distributing ConstructorWithVirtualCallsSpecs.cs among 5 other files     - CreationOptionsSpecs.cs     - CreationSpecs.cs     - FakeScopeSpecs.cs     - MakesVirtualCallInConstructor.cs     - UnconfiguredFakeSpecs.cs
