+++ Source/FakeItEasy.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 2013
-VisualStudioVersion = 12.0.31101.0
+VisualStudioVersion = 12.0.21005.1
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{A6A7B553-5359-4FB9-B19C-8A23004F49DB}"
 	ProjectSection(SolutionItems) = preProject
@@ -64,6 +64,9 @@ EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Integration", "Integration", "{5D41D463-D53B-4ADC-90AD-44AA03D6340D}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Specs", "Specs", "{575A3C3B-17BA-4A8A-9F53-05320913D0A6}"
+	ProjectSection(SolutionItems) = preProject
+		CreationOptionsSpecs.cs = CreationOptionsSpecs.cs
+	EndProjectSection
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FakeItEasy.Win81.Tests", "FakeItEasy.Win81.Tests\FakeItEasy.Win81.Tests.csproj", "{A84169B8-3CAC-415B-A367-A167836B034C}"
 	ProjectSection(ProjectDependencies) = postProject
@@ -138,19 +141,19 @@ Global
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{1ED2F1A2-40E3-498C-B13F-AF066EACF58F} = {F9C87C93-FE80-45E1-8CFF-8D6E451F7939}
-		{0E26272D-045A-47C2-8928-BD9B980115D5} = {DC5EAE30-96E3-4D38-B3C5-33E587C9FB76}
-		{C291B61B-0729-4DC5-B8E4-147A8E3F801F} = {5D41D463-D53B-4ADC-90AD-44AA03D6340D}
-		{69A27B0A-710B-4CEC-B159-56E34C788665} = {5D41D463-D53B-4ADC-90AD-44AA03D6340D}
 		{193C944B-733B-4E01-87A3-73F7AB40CDC7} = {F9C87C93-FE80-45E1-8CFF-8D6E451F7939}
-		{B486F5FE-4371-4E57-BC21-36DCF9F1895A} = {DC5EAE30-96E3-4D38-B3C5-33E587C9FB76}
-		{FA539A2D-98DC-4300-9A2C-EC6FE38A7392} = {575A3C3B-17BA-4A8A-9F53-05320913D0A6}
-		{E029C5AB-EC90-45B0-BBD9-29A60864B137} = {DC5EAE30-96E3-4D38-B3C5-33E587C9FB76}
-		{30396678-D653-4C67-A507-1749424A24CD} = {5D41D463-D53B-4ADC-90AD-44AA03D6340D}
-		{19A3E9ED-D8A6-463A-83E6-D59ABA52F1BF} = {DC5EAE30-96E3-4D38-B3C5-33E587C9FB76}
 		{DC5EAE30-96E3-4D38-B3C5-33E587C9FB76} = {52D18A62-B744-48F6-A4B5-1382F83FA437}
 		{5D41D463-D53B-4ADC-90AD-44AA03D6340D} = {52D18A62-B744-48F6-A4B5-1382F83FA437}
 		{575A3C3B-17BA-4A8A-9F53-05320913D0A6} = {52D18A62-B744-48F6-A4B5-1382F83FA437}
+		{B486F5FE-4371-4E57-BC21-36DCF9F1895A} = {DC5EAE30-96E3-4D38-B3C5-33E587C9FB76}
+		{E029C5AB-EC90-45B0-BBD9-29A60864B137} = {DC5EAE30-96E3-4D38-B3C5-33E587C9FB76}
+		{19A3E9ED-D8A6-463A-83E6-D59ABA52F1BF} = {DC5EAE30-96E3-4D38-B3C5-33E587C9FB76}
+		{0E26272D-045A-47C2-8928-BD9B980115D5} = {DC5EAE30-96E3-4D38-B3C5-33E587C9FB76}
 		{A84169B8-3CAC-415B-A367-A167836B034C} = {DC5EAE30-96E3-4D38-B3C5-33E587C9FB76}
+		{69A27B0A-710B-4CEC-B159-56E34C788665} = {5D41D463-D53B-4ADC-90AD-44AA03D6340D}
+		{30396678-D653-4C67-A507-1749424A24CD} = {5D41D463-D53B-4ADC-90AD-44AA03D6340D}
+		{C291B61B-0729-4DC5-B8E4-147A8E3F801F} = {5D41D463-D53B-4ADC-90AD-44AA03D6340D}
+		{FA539A2D-98DC-4300-9A2C-EC6FE38A7392} = {575A3C3B-17BA-4A8A-9F53-05320913D0A6}
 	EndGlobalSection
 	GlobalSection(MonoDevelopProperties) = preSolution
 		version = 0.1
